// MVCApi.AngularClient/src/api/model/createCustomer.ts
/**
 * MVCApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CreateCustomer { 
    firstName?: string | null;
    lastName?: string | null;
    dateOfBirth?: string;
    country?: string | null;
    city?: string | null;
    street?: string | null;
    streetNumber?: string | null;
    postCode?: string | null;
    email?: string | null;
    phoneNumber?: string | null;
}